Resources + Control implementation
Added a resource helper to help load images. It contains an option to
fallback to a magenta image. If the image is drawn using a width and
height, the magenta image will take up the entire space of the missing
image to help identify this problem to the developer.
Added a control base implementation. If more controls are needed, simply
extend the ControlBase and define what happens when it is clicked and
how it should draw itself.
Added a button implementation to replace the current NewGame button.